Preliminary refactor EBL handling to separate file (src/ebl.py). Also moved required globals to ebl.py and set up handling in __name__.
